requirements.txt
numpy
ma-gym
pygame
matplotlib
scipy
